requirements.txt
alembic==1.16.4
annotated-types==0.7.0
autopep8==2.3.2
blinker==1.9.0
click==8.2.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-Pydantic==0.11.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==22.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycodestyle==2.14.0
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.2
pytest-mock==3.14.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.14.1
Werkzeug==3.0.6
